+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fix
+- Add support for 'target' property for the shortcode [maestro]
+
 ## [1.11.0] 2026-01-07
 ### Add
 - Add the shortcode '[maestro]' to create a link to a master